Broadcom Android Release 8.0 Phase 17.4 (rel 1) code drop

+++ magnum/portinginterface/vce/src/bvce_debug_priv.c
@@ -45,8 +45,6 @@
 #endif
 /* base modules */
 #include "bstd.h"           /* standard types */
-#include "berr.h"           /* error code */
-#include "bdbg.h"           /* debug interface */
 #include "bkni.h"           /* kernel interface */
 
 #include <stdio.h>
@@ -149,7 +147,7 @@ void
 BVCE_Debug_P_WriteLogBuffer_isr(
    BVCE_Debug_Log_Handle hLog,
    const void * pBuffer,
-   size_t uiNumBytes
+   unsigned uiNumBytes
    )
 {
    fwrite( pBuffer, 1, uiNumBytes, hLog->hFile );
